+++ ratect-core/src/lib.rs
@@ -2,3 +2,4 @@ pub mod config;
 pub mod docker;
 pub mod engine;
 pub mod expressions;
+pub mod user;